Refactored logout to sidebar
